Refactor: Modularity in minimal
