+++ app/src/main/java/com/codehanzoom/greenwalk/nav/NavigationGraph.kt
@@ -1,10 +1,22 @@
 package com.codehanzoom.greenwalk.nav
 
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.width
+import androidx.compose.material.BottomNavigationItem
+import androidx.compose.material.Icon
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
+import androidx.navigation.compose.currentBackStackEntryAsState
 import com.codehanzoom.greenwalk.HomeScreen
+import com.codehanzoom.greenwalk.ui.theme.GreenMain
 
 @Composable
 fun NavigationGraph(navController: NavHostController) {
@@ -22,4 +34,50 @@ fun NavigationGraph(navController: NavHostController) {
             //MypageScreen()
         }
     }
+}
+
+@Composable
+fun BottomNavigation(navController: NavHostController) {
+    val items = listOf<BottomNavItem>(
+        BottomNavItem.Home,
+        BottomNavItem.Market,
+        BottomNavItem.Newsfeed,
+        BottomNavItem.Mypage
+    )
+
+    androidx.compose.material.BottomNavigation(
+        backgroundColor = Color.White,
+        contentColor = Color(0xFF3F414E)
+    ) {
+        val navBackStackEntry by navController.currentBackStackEntryAsState()
+        val currentRoute = navBackStackEntry?.destination?.route
+
+        items.forEach { item ->
+            BottomNavigationItem(
+                icon = {
+                    Icon(
+                        painter = painterResource(id = item.icon),
+                        contentDescription = stringResource(id = item.title),
+                        modifier = Modifier
+                            .width(26.dp)
+                            .height(26.dp)
+                    )
+                },
+                selectedContentColor = GreenMain,
+                unselectedContentColor = Color.Gray,
+                selected = currentRoute == item.screenRoute,
+                alwaysShowLabel = false,
+                onClick = {
+                    navController.navigate(item.screenRoute) {
+                        navController.graph.startDestinationRoute?.let {
+                            popUpTo(it) { saveState = true }
+                        }
+                        launchSingleTop = true
+                        restoreState = true
+                    }
+                }
+            )
+        }
+    }
+
 }
